Allow users to optionally add an extra command line switch to the fzf-file issued command
Let's say you are using rg (ripgrep) as the program that will list the file
candidates that should be opened by kakoune. Usually this will ignore all
the files that match .gitignore, if you are in a git folder. *This is usually what you want*

But sometimes you want to be able to open *any* file, regardless of .gitignore

So, add a new parameter to the `fzf-file` that will allow you to pass an arbtrary switch
the fzf-file command that is executed. Here we may want to pass `--no-ignore` to
the rg command, which will cause it to *not* ignore files matched by .gitignore

Here is the sample configuration for your kakrc file so that this all works as you want

```kak
hook global ModuleLoaded fzf-file %{
    # BTW this is slightly different from the default fzf.kak config as it omits the
    # --hidden flag. Usually we _don't_ want to see hidden files like .git/* etc.
    set global fzf_file_command "rg -L --files"
}

hook global ModuleLoaded fzf %{
    map global fzf -docstring "open file in dir, showing even .gitignore-ed files" \
        'o' '<esc>: require-module fzf-file; fzf-file extra-switch=--no-ignore<ret>'
    # BTW, if you want to pass two extra switches you can do something like
    # ...; fzf-file %{extra-switch=--no-ignore --hidden}
}
```

This was just an example. In general, the content of extra-switch can be used to provide
an alternate way of listing files using any of your favorite programs (e.g. rg, ag, find etc.)
by appending a custom command line switch.